# apps/api/src/app/users/dto/avatar-upload-session.dto.ts
import { ApiProperty } from '@nestjs/swagger';

export class AvatarUploadSessionDto {
  @ApiProperty()
  userId: string;

  @ApiProperty()
  storagePath: string;

  @ApiProperty()
  uploadToken: string;
}