# apps/api/src/app/users/entities/user.entity.ts
import {
  Column,
  CreateDateColumn,
  Entity,
  PrimaryGeneratedColumn,
  UpdateDateColumn,
} from 'typeorm';
import { AuthProvider, GenderEnum, UserRole } from '../../../common';
import { ApiProperty } from '@nestjs/swagger';

@Entity('users')
export class User {
  @ApiProperty()
  @PrimaryGeneratedColumn('uuid')
  id: string;

  @ApiProperty()
  @Column({ unique: true })
  email: string;

  @ApiProperty({ enum: AuthProvider })
  @Column({ type: 'enum', enum: AuthProvider, default: AuthProvider.PASSWORD })
  authProvider: AuthProvider;

  @ApiProperty()
  @Column()
  firstName: string;

  @ApiProperty()
  @Column()
  lastName: string;

  @ApiProperty({ required: false })
  @Column({ nullable: true })
  avatar?: string;

  @ApiProperty({ required: false })
  @Column({ type: 'text', nullable: true })
  avatarStoragePath?: string | null;

  @ApiProperty()
  @Column({ type: 'enum', enum: UserRole, default: UserRole.PARTICIPANT })
  role: UserRole;

  @ApiProperty()
  @Column({ default: false })
  isEmailVerified: boolean;

  @ApiProperty({ required: false })
  @Column({ type: 'timestamptz', nullable: true })
  emailVerifiedAt?: Date;

  @ApiProperty({ required: false })
  @Column({ type: 'timestamptz', nullable: true })
  emailVerificationSentAt?: Date;

  @ApiProperty({ required: false })
  @Column({ type: 'timestamptz', nullable: true })
  passwordResetSentAt?: Date;

  @ApiProperty({ required: false })
  @Column({ nullable: true })
  googleId?: string;

  @ApiProperty({ required: false })
  @Column({ nullable: true, unique: true })
  firebaseUid?: string;

  @ApiProperty()
  @Column({ default: true })
  isActive: boolean;

  @ApiProperty()
  @CreateDateColumn({ type: 'timestamptz' })
  createdAt: Date;

  @ApiProperty()
  @UpdateDateColumn({ type: 'timestamptz' })
  updatedAt: Date;

  @ApiProperty({ enum: GenderEnum, description: 'Gender of the user', required: false })
  @Column({ type: 'enum', enum: GenderEnum, nullable: true })
  gender?: GenderEnum;

  @ApiProperty({ required: false })
  @Column({ type: 'date', nullable: true })
  birthday?: Date;

  @ApiProperty({ required: false })
  @Column({ nullable: true, unique: true })
  userHandle?: string;

  @ApiProperty({ required: false })
  @Column({ nullable: true })
  phone?: string;

  @ApiProperty({ description: 'Whether the user has accepted the Terms & Conditions' })
  @Column({ default: false })
  termsAccepted: boolean;

  @ApiProperty({ required: false, description: 'Timestamp when Terms & Conditions were accepted' })
  @Column({ type: 'timestamptz', nullable: true })
  termsAcceptedAt?: Date;

  @ApiProperty({ description: 'Whether the user has accepted the Privacy Policy' })
  @Column({ default: false })
  privacyPolicyAccepted: boolean;

  @ApiProperty({ required: false, description: 'Timestamp when Privacy Policy was accepted' })
  @Column({ type: 'timestamptz', nullable: true })
  privacyPolicyAcceptedAt?: Date;

  @Column({ nullable: true, select: false })
  emailVerificationToken?: string;

  @Column({ type: 'timestamptz', nullable: true })
  emailVerificationTokenExpiresAt?: Date;
}