# packages/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all users",
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad request."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new user",
        "tags": [
          "users"
        ]
      }
    },
    "/users/me": {
      "get": {
        "operationId": "UsersController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the authenticated user — 404 if not in DB",
        "tags": [
          "users"
        ]
      }
    },
    "/users/me/frequent-crew": {
      "get": {
        "operationId": "UsersController_getFrequentCrew",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FrequentCrewDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the frequently seen crew for the authenticated user",
        "tags": [
          "users"
        ]
      }
    },
    "/users/auth-provider-check": {
      "post": {
        "operationId": "UsersController_checkAuthProvider",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckAuthProviderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "exists": true,
                    "authProvider": "password"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Check whether an email already belongs to an existing auth provider",
        "tags": [
          "users"
        ]
      }
    },
    "/users/session-cookie": {
      "post": {
        "operationId": "UsersController_createSessionCookie",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionCookieDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "sessionCookie": "eyJhbGciOiJSUzI1NiIsImtpZCI6Ij...",
                    "expiresIn": 604800
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Exchange a Firebase ID token for a 7-day session cookie",
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "User not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a user by id",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "User not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a user",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "User not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a user",
        "tags": [
          "users"
        ]
      }
    },
    "/users/sync": {
      "post": {
        "operationId": "UsersController_sync",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upsert the authenticated Firebase user into the DB",
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}/avatar/upload-url": {
      "post": {
        "operationId": "UsersController_createAvatarUploadSession",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAvatarUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvatarUploadSessionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create signed upload session for profile avatar",
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}/avatar/complete": {
      "post": {
        "operationId": "UsersController_completeAvatarUpload",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Finalize signed avatar upload and replace previous avatar",
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Documentation",
    "description": "The API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "authProvider": {
            "type": "string",
            "enum": [
              "password",
              "google"
            ]
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "avatar": {
            "type": "string"
          },
          "avatarStoragePath": {
            "type": "object"
          },
          "role": {
            "type": "string"
          },
          "isEmailVerified": {
            "type": "boolean"
          },
          "emailVerifiedAt": {
            "format": "date-time",
            "type": "string"
          },
          "emailVerificationSentAt": {
            "format": "date-time",
            "type": "string"
          },
          "passwordResetSentAt": {
            "format": "date-time",
            "type": "string"
          },
          "googleId": {
            "type": "string"
          },
          "firebaseUid": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female",
              "other"
            ],
            "description": "Gender of the user"
          },
          "birthday": {
            "format": "date-time",
            "type": "string"
          },
          "userHandle": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "termsAccepted": {
            "type": "boolean",
            "description": "Whether the user has accepted the Terms & Conditions"
          },
          "termsAcceptedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp when Terms & Conditions were accepted"
          },
          "privacyPolicyAccepted": {
            "type": "boolean",
            "description": "Whether the user has accepted the Privacy Policy"
          },
          "privacyPolicyAcceptedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp when Privacy Policy was accepted"
          }
        },
        "required": [
          "id",
          "email",
          "authProvider",
          "firstName",
          "lastName",
          "role",
          "isEmailVerified",
          "isActive",
          "createdAt",
          "updatedAt",
          "termsAccepted",
          "privacyPolicyAccepted"
        ]
      },
      "UserProfileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "authProvider": {
            "type": "string",
            "enum": [
              "password",
              "google"
            ]
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "avatar": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "participant"
            ]
          },
          "phone": {
            "type": "string"
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female",
              "other"
            ]
          },
          "birthday": {
            "format": "date-time",
            "type": "string",
            "example": "1990-01-01"
          },
          "userHandle": {
            "type": "string"
          },
          "isEmailVerified": {
            "type": "boolean"
          },
          "isActive": {
            "type": "boolean"
          },
          "dropCount": {
            "type": "number",
            "description": "Number of drops created by this user"
          },
          "crewReached": {
            "type": "number",
            "description": "Total unique crew members reached across all drops"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "authProvider",
          "firstName",
          "lastName",
          "role",
          "isEmailVerified",
          "isActive",
          "dropCount",
          "crewReached",
          "createdAt",
          "updatedAt"
        ]
      },
      "FrequentCrewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "avatar": {
            "type": "string"
          },
          "userHandle": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "frequencyCount": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "createdAt",
          "frequencyCount"
        ]
      },
      "CheckAuthProviderDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "CreateSessionCookieDto": {
        "type": "object",
        "properties": {
          "idToken": {
            "type": "string",
            "description": "Firebase ID token from the authenticated client session.",
            "example": "eyJhbGciOiJSUzI1NiIsImtpZCI6Ij..."
          }
        },
        "required": [
          "idToken"
        ]
      },
      "SyncUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "John",
            "maxLength": 100
          },
          "lastName": {
            "type": "string",
            "example": "Doe",
            "maxLength": 100
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female",
              "other"
            ]
          },
          "birthday": {
            "type": "string",
            "example": "1990-01-01"
          },
          "userHandle": {
            "type": "string",
            "example": "jane_doe"
          },
          "authMode": {
            "type": "string",
            "enum": [
              "login",
              "signup"
            ]
          },
          "authProvider": {
            "type": "string",
            "enum": [
              "password",
              "google"
            ]
          },
          "email": {
            "type": "string",
            "example": "user@example.com"
          }
        }
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "authProvider": {
            "type": "string",
            "enum": [
              "password",
              "google"
            ],
            "default": "password"
          },
          "firstName": {
            "type": "string",
            "example": "Jane"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "firebaseUid": {
            "type": "string",
            "example": "firebase-uid-123"
          },
          "avatar": {
            "type": "string",
            "example": "https://example.com/avatar.jpg"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "participant"
            ],
            "default": "participant"
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female",
              "other"
            ]
          },
          "birthday": {
            "format": "date-time",
            "type": "string",
            "example": "1990-01-01"
          },
          "userHandle": {
            "type": "string",
            "example": "jane_doe"
          },
          "phone": {
            "type": "string",
            "example": "+639123456789"
          },
          "termsAccepted": {
            "type": "boolean",
            "default": false
          },
          "termsAcceptedAt": {
            "format": "date-time",
            "type": "string"
          },
          "privacyPolicyAccepted": {
            "type": "boolean",
            "default": false
          },
          "privacyPolicyAcceptedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "email",
          "firstName",
          "lastName"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "Jane"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "avatar": {
            "type": "string",
            "example": "https://example.com/avatar.jpg"
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female",
              "other"
            ]
          },
          "birthday": {
            "format": "date-time",
            "type": "string",
            "example": "1990-01-01"
          },
          "userHandle": {
            "type": "string",
            "example": "jane_doe"
          },
          "phone": {
            "type": "string",
            "example": "+639123456789"
          },
          "termsAccepted": {
            "type": "boolean",
            "default": false
          },
          "termsAcceptedAt": {
            "format": "date-time",
            "type": "string"
          },
          "privacyPolicyAccepted": {
            "type": "boolean",
            "default": false
          },
          "privacyPolicyAcceptedAt": {
            "format": "date-time",
            "type": "string"
          }
        }
      },
      "CreateAvatarUploadDto": {
        "type": "object",
        "properties": {
          "mimeType": {
            "type": "string",
            "enum": [
              "image/jpeg",
              "image/jpg",
              "image/png"
            ]
          },
          "sizeBytes": {
            "type": "number",
            "description": "Original file size in bytes",
            "maximum": 5242880
          },
          "width": {
            "type": "number"
          },
          "height": {
            "type": "number"
          }
        },
        "required": [
          "mimeType",
          "sizeBytes"
        ]
      },
      "AvatarUploadSessionDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "storagePath": {
            "type": "string"
          },
          "uploadToken": {
            "type": "string"
          }
        },
        "required": [
          "userId",
          "storagePath",
          "uploadToken"
        ]
      },
...
    }
  }
}